Patch non-segwit P2SH scriptSig for replay-protection
The nonstandard script-type was added specifically for the 
OP_CHECKDATASIGVERIFY script template used for BCH/BSV 
replay-protection. In order to support this script type, we must handle
the scriptSig differently than regular P2SH.

In the future, maybe this template should be built outside of the lib
using the Payments API directly, rather than implementing support for it
in the lib.